+++ CHANGELOG.md
@@ -1,9 +1,23 @@
 # Changelog
 
+## 1.2.5 - 2026-06-26
+- Added gender marker to employee personal information
+- Added employee ID field to employee personal information
+- Updated dropdowns in New Employee dialog to use Select component
+- Removed some unused state
+- Added ability to cancel onboarding for employees (this will remove the employee from view, but preserve all data)
+- Fixed bug where reattempting quiz after failure would push results twice
+- Added cancelled onboarding count to dashboard
+- Removed time tracking for bloodborne pathogens popup
+- Added logging to audit logs for when caregivers/employees complete a step
+- Added field to support storing expiration dates for caregiver documents in the database
+
+
 ## 1.2.4 - 2026-06-14
 - Fixed bug where start date on table appeared one day behind
 - Fixed bug where start date on offer letter would appear one day behind
 - Added reauthentication requirement when viewing Social Security Card / Other I-9 Supporting Documentation
+
 
 ## 1.2.3 - 2026-06-13
 - Added password verification before accessing tax documents
@@ -12,6 +26,7 @@
 - Added position of admin to I-9 field
 - Reorganized ordering of steps
 - Added new step, "How to Use SurePayroll"
+
 
 ## 1.2.2 - 2026-06-13